backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.3
